chore: address issue #607
